Took pelletier 54 offline.
Getting rid of preprocessing makes proving pelletier 54 take a lot
longer, too long for it to be a proper test. This is an instance of a
more general problem, namely that proof time depends quite a lot on the
initial order of clauses. Some canonical order would probably help with
this.